+++ src/main/kotlin/com/mintlify/document/ui/MyToolWindowFactory.kt
@@ -1,5 +1,5 @@
 // Copyright 2000-2022 JetBrains s.r.o. and other contributors. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
-package com.mintlify.document.extensions
+package com.mintlify.document.ui
 
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.wm.ToolWindow
