+++ src/main/resources/META-INF/plugin.xml
@@ -9,8 +9,8 @@
     <extensions defaultExtensionNs="com.intellij">
         <applicationService serviceImplementation="com.mintlify.document.services.MyApplicationService"/>
         <projectService serviceImplementation="com.mintlify.document.services.MyProjectService"/>
-        <toolWindow id="Sample Calendar" secondary="true" icon="AllIcons.General.Modified" anchor="right"
-                    factoryClass="com.mintlify.document.extensions.MyToolWindowFactory"/>
+        <toolWindow id="AI Doc Writer" icon="AllIcons.General.Modified" anchor="right"
+                    factoryClass="com.mintlify.document.ui.MyToolWindowFactory"/>
     </extensions>
 
     <applicationListeners>
